Renderer: Switch num_bars to uint

+++ shaders/bars/x_gradient.glsl
@@ -3,7 +3,7 @@
 uniform float[120] heights;
 uniform float width;
 uniform float height;
-uniform float num_bars;
+uniform uint num_bars;
 
 void main() {
     vec3 color1 = vec3(69.f, 104.f, 220.f);
